+++ src/main/java/com/junak/scorekeeper/entity/PlayerFieldingDetails.java
@@ -26,7 +26,7 @@ public class PlayerFieldingDetails {
     private int gamesPlayed;
 
     @Column(name = "INN")
-    private int innings;
+    private double innings;
 
     @Column(name = "TC")
     private int totalChances;
@@ -88,11 +88,11 @@ public class PlayerFieldingDetails {
         this.gamesPlayed = gamesPlayed;
     }
 
-    public int getInnings() {
+    public double getInnings() {
         return innings;
     }
 
-    public void setInnings(int innings) {
+    public void setInnings(double innings) {
         this.innings = innings;
     }
